fixing a bunch of warnings in visual studio 2010

+++ Testing/Source/DataStructureAndEncodingDefinition/Cxx/TestDS.cxx
@@ -77,7 +77,7 @@ std::string to_string ( Float data ) {
 #endif
             }
             else if( ldata < 0)
-                digits += ldata; // (zeros before first significant digit)
+                digits += (unsigned long)ldata; // (zeros before first significant digit)
         }
     }
     /*
@@ -424,7 +424,7 @@ int TestDS(int, char *[])
     for (int i = 0; i<random_count; i++)
     {
         // Create something that looks like a random double
-        int rand_exp =  ( ( std::rand() / (double)(RAND_MAX) ) * (max_exp - min_exp) ) + min_exp;
+        double rand_exp =  ( ( std::rand() / (double)(RAND_MAX) ) * (max_exp - min_exp) ) + min_exp;
         double rand = static_cast<double>(std::rand()) * pow(10., rand_exp);
         if (rand != rand) {i--; continue;} // nan
         if (rand == std::numeric_limits<double>::infinity()) {i--; continue;} // inf
